Github URL building w/ tests coverage
