+++ Classes/Repository/RepositoryCodeBlobViewController.swift
@@ -21,13 +21,7 @@ final class RepositoryCodeBlobViewController: UIViewController, EmptyViewDelegat
     private let emptyView = EmptyView()
     private var sharingPayload: Any?
     private var repoUrl: URL? {
-        let builder = URLBuilder.github()
-            .add(path: repo.owner)
-            .add(path: repo.name)
-            .add(path: "blob")
-            .add(path: branch)
-        path.components.forEach { builder.add(path: $0) }
-        return builder.url
+        return GithubURL.codeBlob(repo: repo, branch: branch, path: path)
     }
 
     private lazy var moreOptionsItem: UIBarButtonItem = {
@@ -81,7 +75,7 @@ final class RepositoryCodeBlobViewController: UIViewController, EmptyViewDelegat
 
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        if let url = externalURL {
+        if let url = repoUrl {
             setupUserActivity(with: HandoffInformator(
                 activityName: "viewCodeBlob",
                 activityTitle: "\(repo.owner)/\(repo.name)/\(branch)/" + path.components.joined(separator: "/"),
@@ -109,10 +103,6 @@ final class RepositoryCodeBlobViewController: UIViewController, EmptyViewDelegat
     }
 
     // MARK: Private API
-
-    private var externalURL: URL? {
-        return URLBuilder.github().add(paths: [repo.owner, repo.name, "blob", branch] + path.components).url
-    }
 
     @objc func onFileNavigationTitle(sender: UIView) {
         showAlert(filePath: path, sender: sender)
